JPERF-1079 Extract StoppableNode class outside StartedNode.
