+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/jira/StartedNode.kt
@@ -5,10 +5,7 @@ import com.atlassian.performance.tools.awsinfrastructure.api.aws.AwsCli
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraGcLog
 import com.atlassian.performance.tools.infrastructure.api.process.RemoteMonitoringProcess
 import com.atlassian.performance.tools.ssh.api.Ssh
-import com.atlassian.performance.tools.ssh.api.SshHost
 import java.time.Duration
-import javax.json.Json
-import javax.json.JsonObject
 
 class StartedNode(
     private val name: String,
@@ -77,53 +74,5 @@ class StartedNode(
 
     override fun toString() = name
 
-    fun toStoppableNode(): StoppableNode {
-        return StoppableNode.Builder(
-            jiraPath = jiraPath,
-            ssh = ssh
-        ).build()
-    }
-
-    class StoppableNode private constructor(
-        private val jiraPath: String,
-        private val ssh: Ssh
-    ) {
-        fun stopNode() {
-            ssh.newConnection().use { ssh ->
-                ssh.execute(
-                    """
-                    |source ~/.profile
-                    |${jiraPath}/bin/stop-jira.sh
-                    """.trimMargin(),
-                    Duration.ofMinutes(3)
-                )
-            }
-        }
-
-        fun toJson(): JsonObject {
-            return Json.createObjectBuilder()
-                .add("jiraPath", jiraPath)
-                .add("ssh", ssh.host.toJson())
-                .build()
-        }
-
-        class Builder(
-            private var jiraPath: String,
-            private var ssh: Ssh
-        ) {
-            constructor(json: JsonObject) : this(
-                jiraPath = json.getString("jiraPath"),
-                ssh = Ssh(
-                    SshHost(json.getJsonObject("ssh"))
-                )
-            )
-
-            fun jiraPath(jiraPath: String) = apply { this.jiraPath = jiraPath }
-            fun ssh(ssh: Ssh) = apply { this.ssh = ssh }
-            fun build() = StoppableNode(
-                jiraPath = jiraPath,
-                ssh = ssh
-            )
-        }
-    }
+    fun toStoppableNode(): StoppableNode = StoppableNode.Builder(jiraPath, ssh).build()
 }
